对于气压计修改

+++ User/Receive.h
@@ -5,6 +5,6 @@ extern int num;
 extern float sum;
 
 char ReceiveVis(void);
-float ReceiveNum(void);
+uint8_t ReceiveNum(void);
 uint8_t ReceiveNum_Gettarget(void);
 #endif
